+++ build-resources/c420ui-rs/src/commands/mod.rs
@@ -2,5 +2,8 @@ pub mod check_command;
 pub mod check_host_dependencies;
 pub mod check_node;
 pub mod doctor;
+pub mod fix_permissions;
 pub mod host_info;
+pub mod remove_paths;
 pub mod run_process;
+pub mod sudo_validate;